[FEATURE] initial quickly checklist screen
