+++ app/src/main/java/wottrich/github/io/androidsmartchecklist/di/AppModule.kt
@@ -5,6 +5,7 @@ import github.io.wottrich.newchecklist.di.newChecklistModule
 import github.io.wottrich.ui.support.di.supportModule
 import wottrich.github.io.datasource.injection.databaseModule
 import wottrich.github.io.impl.di.taskModule
+import wottrich.github.io.quicklychecklist.impl.di.quicklyChecklistModule
 import wottrich.github.io.tools.injection.toolsDispatcherModule
 
 /**
@@ -29,6 +30,7 @@ object AppModule {
         taskModule,
         supportModule,
         newChecklistModule,
+        quicklyChecklistModule,
 
         //Database
         databaseModule,
