+++ app/src/main/java/wottrich/github/io/androidsmartchecklist/presentation/viewmodel/ChecklistSettingsViewModel.kt
@@ -4,8 +4,8 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import wottrich.github.io.androidsmartchecklist.presentation.viewmodel.ChecklistSettingsAllTasksAction.CHECK_ALL
 import wottrich.github.io.androidsmartchecklist.presentation.viewmodel.ChecklistSettingsAllTasksAction.UNCHECK_ALL
-import wottrich.github.io.impl.domain.usecase.ChangeTasksCompletedStatusUseCase
-import wottrich.github.io.impl.domain.usecase.GetTasksUseCase
+import wottrich.github.io.quicklychecklist.impl.domain.usecase.ChangeTasksCompletedStatusUseCase
+import wottrich.github.io.quicklychecklist.impl.domain.usecase.GetTasksUseCase
 import wottrich.github.io.tools.SingleShotEventBus
 import wottrich.github.io.tools.base.BaseViewModel
 import wottrich.github.io.tools.dispatcher.DispatchersProviders
