+++ features/quicklychecklist/impl/build.gradle.kts
@@ -1,0 +1,54 @@
+plugins {
+    id(Plugins.androidLibrary)
+    id(Plugins.kotlinAndroid)
+}
+
+android {
+    compileSdk = AndroidSdk.targetSdk
+    buildToolsVersion = AndroidSdk.buildToolsVersion
+
+    defaultConfig {
+        minSdk = AndroidSdk.minSdk
+        targetSdk = AndroidSdk.targetSdk
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles("consumer-rules.pro")
+    }
+
+    buildTypes {
+        getByName("release") {
+            isMinifyEnabled = false
+            proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
+        }
+    }
+    compileOptions {
+        sourceCompatibility = AndroidSdk.javaVersion
+        targetCompatibility = AndroidSdk.javaVersion
+    }
+    kotlinOptions {
+        jvmTarget = AndroidSdk.javaVersion.toString()
+    }
+    buildFeatures {
+        compose = true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion = Versions.composeVersion
+    }
+}
+
+dependencies {
+
+    kotlinAndCoreKtx()
+
+    implementation(project(path = ":baseui"))
+    implementation(project(path = ":tools"))
+
+    moduleCommonUiCompose()
+
+    composeUi()
+    koin()
+
+    navigation(withAnimation = true)
+
+    unitTest()
+    instrumentalTest()
+}
