+++ features/quicklychecklist/impl/src/androidTest/java/wottrich/github/io/quicklychecklist/impl/ExampleInstrumentedTest.kt
@@ -1,0 +1,24 @@
+package wottrich.github.io.quicklychecklist.impl
+
+import androidx.test.platform.app.InstrumentationRegistry
+import androidx.test.ext.junit.runners.AndroidJUnit4
+
+import org.junit.Test
+import org.junit.runner.RunWith
+
+import org.junit.Assert.*
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        assertEquals("wottrich.github.io.quicklychecklist.impl.test", appContext.packageName)
+    }
+}
