+++ features/task/impl/src/main/java/wottrich/github/io/impl/di/TaskInjection.kt
@@ -2,11 +2,11 @@ package wottrich.github.io.impl.di
 
 import org.koin.core.module.Module
 import org.koin.dsl.module
-import wottrich.github.io.impl.domain.usecase.ChangeTasksCompletedStatusUseCase
-import wottrich.github.io.impl.domain.usecase.GetAddTaskUseCase
-import wottrich.github.io.impl.domain.usecase.GetChangeTaskStatusUseCase
-import wottrich.github.io.impl.domain.usecase.GetDeleteTaskUseCase
-import wottrich.github.io.impl.domain.usecase.GetTasksUseCase
+import wottrich.github.io.quicklychecklist.impl.domain.usecase.ChangeTasksCompletedStatusUseCase
+import wottrich.github.io.quicklychecklist.impl.domain.usecase.GetAddTaskUseCase
+import wottrich.github.io.quicklychecklist.impl.domain.usecase.GetChangeTaskStatusUseCase
+import wottrich.github.io.quicklychecklist.impl.domain.usecase.GetDeleteTaskUseCase
+import wottrich.github.io.quicklychecklist.impl.domain.usecase.GetTasksUseCase
 
 val taskModule = module {
     injectUseCases()
