--- features/task/impl/src/main/java/wottrich/github/io/impl/presentation/ui/TaskRippleLayer.kt
@@ -1,14 +1,0 @@
-package wottrich.github.io.impl.presentation.ui
-
-import androidx.compose.material.ripple.LocalRippleTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import github.io.wottrich.common.ui.compose.ripple.RippleStatusColor
-
-@Composable
-internal fun TaskRippleLayer(isTaskComplete: Boolean, content: @Composable () -> Unit) {
-    CompositionLocalProvider(
-        LocalRippleTheme provides RippleStatusColor(!isTaskComplete),
-        content = content
-    )
-}
