+++ settings.gradle
@@ -13,3 +13,4 @@ include ':ui-support'
 include ':features:newchecklist:impl'
 include ':features:newchecklist:public'
 include ':common-ui-compose'
+include ':features:quicklychecklist:impl'
