+++ tools/src/main/res/drawable/ic_quickly_checklist_24.xml
@@ -1,0 +1,6 @@
+<vector android:autoMirrored="true" android:height="24dp"
+    android:tint="#000000" android:viewportHeight="24"
+    android:viewportWidth="24" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white"
+        android:fillType="evenOdd" android:pathData="M20,3H4C2.9,3 2,3.9 2,5v14c0,1.1 0.9,2 2,2h16c1.1,0 2,-0.9 2,-2V5C22,3.9 21.1,3 20,3zM9,17H6c-0.55,0 -1,-0.45 -1,-1c0,-0.55 0.45,-1 1,-1h3c0.55,0 1,0.45 1,1C10,16.55 9.55,17 9,17zM9,13H6c-0.55,0 -1,-0.45 -1,-1c0,-0.55 0.45,-1 1,-1h3c0.55,0 1,0.45 1,1C10,12.55 9.55,13 9,13zM9,9H6C5.45,9 5,8.55 5,8c0,-0.55 0.45,-1 1,-1h3c0.55,0 1,0.45 1,1C10,8.55 9.55,9 9,9zM18.7,11.12l-3.17,3.17c-0.39,0.39 -1.03,0.39 -1.42,0l-1.41,-1.42c-0.39,-0.39 -0.39,-1.02 0,-1.41c0.39,-0.39 1.02,-0.39 1.41,0l0.71,0.71l2.47,-2.47c0.39,-0.39 1.02,-0.39 1.41,0l0.01,0.01C19.09,10.1 19.09,10.74 18.7,11.12z"/>
+</vector>
